refactor: extract open_in_editor logic into a common function
