+++ src/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod add;
 pub mod edit;
+pub mod editor;
 pub mod env;
 pub mod init;
 pub mod list;
